Fix txpool metrics on re-enabling
